+++ futures/src/backend/wasm.rs
@@ -1,0 +1,2 @@
+//! Backends that are only available on Wasm targets.
+pub mod wasm_bindgen;
